+++ draft-js-mention-plugin/CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Migrate styles to linaria
 - Hide internals in single bundle
 - Add esm support
+- Use lodash-es in esm bundle
 
 ## 3.1.5